backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.0,<5.0
dj-database-url==0.5.0
psycopg2-binary>=2.8,<3.0
python-dotenv>=0.19,<1.0
gunicorn>=20.0,<21.0
django-cors-headers>=3.10,<4.0